LPS-165429 Inline BaseSearchEngine
