+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/ElasticsearchSearchEngine.java
@@ -20,13 +20,11 @@ import com.liferay.portal.events.StartupHelperUtil;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.BaseSearchEngine;
 import com.liferay.portal.kernel.search.IndexSearcher;
 import com.liferay.portal.kernel.search.IndexWriter;
 import com.liferay.portal.kernel.search.SearchEngine;
 import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.PortalRunMode;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Time;
@@ -82,7 +80,7 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
 	immediate = true, property = "search.engine.impl=Elasticsearch",
 	service = {ElasticsearchSearchEngine.class, SearchEngine.class}
 )
-public class ElasticsearchSearchEngine extends BaseSearchEngine {
+public class ElasticsearchSearchEngine implements SearchEngine {
 
 	@Override
 	public synchronized String backup(long companyId, String backupName)
@@ -116,9 +114,22 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 	}
 
 	@Override
-	public void initialize(long companyId) {
-		super.initialize(companyId);
+	public IndexSearcher getIndexSearcher() {
+		return _indexSearcher;
+	}
 
+	@Override
+	public IndexWriter getIndexWriter() {
+		return _indexWriter;
+	}
+
+	@Override
+	public String getVendor() {
+		return "Elasticsearch";
+	}
+
+	@Override
+	public void initialize(long companyId) {
 		_waitForYellowStatus();
 
 		RestHighLevelClient restHighLevelClient =
@@ -153,8 +164,6 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 
 	@Override
 	public void removeCompany(long companyId) {
-		super.removeCompany(companyId);
-
 		CrossClusterReplicationHelper crossClusterReplicationHelper =
 			_crossClusterReplicationHelper;
 
@@ -210,18 +219,6 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 		_waitForYellowStatus();
 	}
 
-	@Override
-	@Reference(target = "(search.engine.impl=Elasticsearch)", unbind = "-")
-	public void setIndexSearcher(IndexSearcher indexSearcher) {
-		super.setIndexSearcher(indexSearcher);
-	}
-
-	@Override
-	@Reference(target = "(search.engine.impl=Elasticsearch)", unbind = "-")
-	public void setIndexWriter(IndexWriter indexWriter) {
-		super.setIndexWriter(indexWriter);
-	}
-
 	public void unsetElasticsearchConnectionManager(
 		ElasticsearchConnectionManager elasticsearchConnectionManager) {
 
@@ -235,8 +232,6 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 	@Activate
 	protected void activate(Map<String, Object> properties) {
 		_checkNodeVersions();
-
-		setVendor(MapUtil.getString(properties, "search.engine.impl"));
 
 		if (StartupHelperUtil.isDBNew()) {
 			for (long companyId : _getIndexedCompanyIds()) {
@@ -479,6 +474,13 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 
 	private IndexFactory _indexFactory;
 	private IndexNameBuilder _indexNameBuilder;
+
+	@Reference(target = "(search.engine.impl=Elasticsearch)")
+	private IndexSearcher _indexSearcher;
+
+	@Reference(target = "(search.engine.impl=Elasticsearch)")
+	private IndexWriter _indexWriter;
+
 	private SearchEngineAdapter _searchEngineAdapter;
 	private SearchEngineInformation _searchEngineInformation;
