--- portal-kernel/src/com/liferay/portal/kernel/search/BaseSearchEngine.java
@@ -1,92 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.kernel.search;
-
-import com.liferay.portal.kernel.search.dummy.DummyIndexSearcher;
-import com.liferay.portal.kernel.search.dummy.DummyIndexWriter;
-
-/**
- * @author Bruno Farache
- * @author Carlos Sierra Andrés
- * @author Marcellus Tavares
- */
-public class BaseSearchEngine implements SearchEngine {
-
-	/**
-	 * @throws SearchException
-	 */
-	@Override
-	public String backup(long companyId, String backupName)
-		throws SearchException {
-
-		return null;
-	}
-
-	@Override
-	public IndexSearcher getIndexSearcher() {
-		return _indexSearcher;
-	}
-
-	@Override
-	public IndexWriter getIndexWriter() {
-		return _indexWriter;
-	}
-
-	@Override
-	public String getVendor() {
-		return _vendor;
-	}
-
-	@Override
-	public void initialize(long companyId) {
-	}
-
-	/**
-	 * @throws SearchException
-	 */
-	@Override
-	public void removeBackup(long companyId, String backupName)
-		throws SearchException {
-	}
-
-	@Override
-	public void removeCompany(long companyId) {
-	}
-
-	/**
-	 * @throws SearchException
-	 */
-	@Override
-	public void restore(long companyId, String backupName)
-		throws SearchException {
-	}
-
-	public void setIndexSearcher(IndexSearcher indexSearcher) {
-		_indexSearcher = indexSearcher;
-	}
-
-	public void setIndexWriter(IndexWriter indexWriter) {
-		_indexWriter = indexWriter;
-	}
-
-	public void setVendor(String vendor) {
-		_vendor = vendor;
-	}
-
-	private IndexSearcher _indexSearcher = new DummyIndexSearcher();
-	private IndexWriter _indexWriter = new DummyIndexWriter();
-	private String _vendor;
-
-}
